Add ability to update and organize.
When new files are added to the directory, 'ptag update' can be
used to find them and add tags to them. The new 'ptag organize'
command can be used to sort the entries in the database. This is
all quite messy at the moment and could use some cleaning up.